Suggestion for modules organisation
