+++ crabs/detector/models.py
@@ -8,7 +8,7 @@ from torchvision.models.detection import (
     fasterrcnn_resnet50_fpn_v2,
 )
 
-from crabs.detection.evaluate_utils import compute_confusion_matrix_elements
+from crabs.detector.utils.evaluate import compute_confusion_matrix_elements
 
 
 class FasterRCNN(LightningModule):
