+++ crabs/detector/train_model.py
@@ -9,19 +9,19 @@ import torch
 import yaml  # type: ignore
 from lightning.pytorch.callbacks import ModelCheckpoint
 
-from crabs.detection.datamodules import CrabsDataModule
-from crabs.detection.detection_utils import (
-    get_checkpoint_type,
-    log_data_augm_as_artifacts,
+from crabs.detector.datamodules import CrabsDataModule
+from crabs.detector.models import FasterRCNN
+from crabs.detector.utils.detection import (
     prep_annotation_files,
     prep_img_directories,
     set_mlflow_run_name,
     setup_mlflow_logger,
     slurm_logs_as_artifacts,
 )
-from crabs.detection.models import FasterRCNN
-from crabs.detection.optuna_utils import (
-    compute_optimal_hyperparameters,
+from crabs.detector.utils.optuna import compute_optimal_hyperparameters
+from crabs.detector.utils.train import (
+    get_checkpoint_type,
+    log_data_augm_as_artifacts,
 )
