+++ tests/test_unit/test_train_model.py
@@ -2,11 +2,9 @@ from pathlib import Path
 
 import pytest
 
-from crabs.detection.detection_utils import (
+from crabs.detector.train_model import train_parse_args
+from crabs.detector.utils.detection import (
     DEFAULT_ANNOTATIONS_FILENAME as DEFAULT_ANNOT_FILENAME,
-)
-from crabs.detection.train_model import (
-    train_parse_args,
 )
 
 DATASET_1 = "/home/data/dataset1"
@@ -23,7 +21,7 @@ def test_prep_img_directories(dataset_dirs: list):
     """
     Test parsing of image directories when training a model.
     """
-    from crabs.detection.train_model import DectectorTrain
+    from crabs.detector.train_model import DectectorTrain
 
     # prepare parser
     train_args = train_parse_args(["--dataset_dirs"] + dataset_dirs)
@@ -52,7 +50,7 @@ def test_prep_annotation_files_single_dataset(annotation_files, expected):
     Test parsing of annotation files when training a model on a single dataset.
     """
 
-    from crabs.detection.train_model import DectectorTrain
+    from crabs.detector.train_model import DectectorTrain
 
     # prepare CLI arguments
     cli_inputs = ["--dataset_dirs", DATASET_1]
@@ -95,7 +93,7 @@ def test_prep_annotation_files_multiple_datasets(annotation_files, expected):
     Test parsing of annotation files when training a model on two datasets.
     """
 
-    from crabs.detection.train_model import DectectorTrain
+    from crabs.detector.train_model import DectectorTrain
 
     # prepare CLI arguments considering multiple dataset
     cli_inputs = ["--dataset_dirs", DATASET_1, DATASET_2]
